+++ src/appmixer/pinterest/core/CreatePin/component.json
@@ -1,7 +1,7 @@
 {
-    "name": "appmixer.pinterest.core.CreateaBoard",
+    "name": "appmixer.pinterest.core.CreatePin",
     "author": "Appmixer <info@appmixer.com>",
-    "description": "Creates a new board on Pinterest.",
+    "description": "Creates a new pin on a specified board.",
     "version": "1.0.0",
     "private": false,
     "auth": {
@@ -17,27 +17,45 @@
             "schema": {
                 "type": "object",
                 "properties": {
-                    "name": {
+                    "boardId": {
+                        "type": "string"
+                    },
+                    "title": {
                         "type": "string"
                     },
                     "description": {
+                        "type": "string"
+                    },
+                    "image_url": {
                         "type": "string"
                     }
                 }
             },
             "inspector": {
                 "inputs": {
-                    "name": {
+                    "board_id": {
                         "type": "text",
-                        "tooltip": "The name of the board.",
-                        "label": "Name",
+                        "tooltip": "The ID of the board where the pin will be created.",
+                        "label": "Board Id",
                         "index": 0
+                    },
+                    "title": {
+                        "type": "text",
+                        "tooltip": "The title of the pin.",
+                        "label": "Title",
+                        "index": 1
                     },
                     "description": {
                         "type": "text",
-                        "tooltip": "A description for the board.",
+                        "tooltip": "A description for the pin.",
                         "label": "Description",
-                        "index": 1
+                        "index": 2
+                    },
+                    "image_url": {
+                        "type": "text",
+                        "tooltip": "The URL of the image for the pin.",
+                        "label": "Image Url",
+                        "index": 3
                     }
                 }
             }
@@ -55,15 +73,15 @@
                     }
                 },
                 {
-                    "label": "Name",
-                    "value": "name",
+                    "label": "Title",
+                    "value": "title",
                     "schema": {
                         "type": "string"
                     }
                 },
                 {
-                    "label": "Description",
-                    "value": "description",
+                    "label": "Board Id",
+                    "value": "board_id",
                     "schema": {
                         "type": "string"
                     }
