--- src/appmixer/pinterest/core/CreateaPin/component.json
@@ -1,93 +1,0 @@
-{
-    "name": "appmixer.pinterest.core.CreateaPin",
-    "author": "Appmixer <info@appmixer.com>",
-    "description": "Creates a new pin on a specified board.",
-    "version": "1.0.0",
-    "private": false,
-    "auth": {
-        "service": "appmixer:pinterest"
-    },
-    "quota": {
-        "manager": "appmixer:pinterest",
-        "resources": "requests"
-    },
-    "inPorts": [
-        {
-            "name": "in",
-            "schema": {
-                "type": "object",
-                "properties": {
-                    "board_id": {
-                        "type": "string"
-                    },
-                    "title": {
-                        "type": "string"
-                    },
-                    "description": {
-                        "type": "string"
-                    },
-                    "image_url": {
-                        "type": "string"
-                    }
-                }
-            },
-            "inspector": {
-                "inputs": {
-                    "board_id": {
-                        "type": "text",
-                        "tooltip": "The ID of the board where the pin will be created.",
-                        "label": "Board Id",
-                        "index": 0
-                    },
-                    "title": {
-                        "type": "text",
-                        "tooltip": "The title of the pin.",
-                        "label": "Title",
-                        "index": 1
-                    },
-                    "description": {
-                        "type": "text",
-                        "tooltip": "A description for the pin.",
-                        "label": "Description",
-                        "index": 2
-                    },
-                    "image_url": {
-                        "type": "text",
-                        "tooltip": "The URL of the image for the pin.",
-                        "label": "Image Url",
-                        "index": 3
-                    }
-                }
-            }
-        }
-    ],
-    "outPorts": [
-        {
-            "name": "out",
-            "options": [
-                {
-                    "label": "Id",
-                    "value": "id",
-                    "schema": {
-                        "type": "string"
-                    }
-                },
-                {
-                    "label": "Title",
-                    "value": "title",
-                    "schema": {
-                        "type": "string"
-                    }
-                },
-                {
-                    "label": "Board Id",
-                    "value": "board_id",
-                    "schema": {
-                        "type": "string"
-                    }
-                }
-            ]
-        }
-    ],
-    "icon": "data:image/svg+xml;base64,PD94bW"
-}
