--- src/appmixer/pinterest/core/DeleteaBoard/component.json
@@ -1,59 +1,0 @@
-{
-    "name": "appmixer.pinterest.core.DeleteaBoard",
-    "author": "Appmixer <info@appmixer.com>",
-    "description": "Deletes a specific board.",
-    "version": "1.0.0",
-    "private": false,
-    "auth": {
-        "service": "appmixer:pinterest"
-    },
-    "quota": {
-        "manager": "appmixer:pinterest",
-        "resources": "requests"
-    },
-    "inPorts": [
-        {
-            "name": "in",
-            "schema": {
-                "type": "object",
-                "properties": {
-                    "board_id": {
-                        "type": "string"
-                    }
-                }
-            },
-            "inspector": {
-                "inputs": {
-                    "board_id": {
-                        "type": "text",
-                        "tooltip": "The ID of the board to delete.",
-                        "label": "Board Id",
-                        "index": 0
-                    }
-                }
-            }
-        }
-    ],
-    "outPorts": [
-        {
-            "name": "out",
-            "options": [
-                {
-                    "label": "Status",
-                    "value": "status",
-                    "schema": {
-                        "type": "string"
-                    }
-                },
-                {
-                    "label": "Message",
-                    "value": "message",
-                    "schema": {
-                        "type": "string"
-                    }
-                }
-            ]
-        }
-    ],
-    "icon": "data:image/svg+xml;base64,PD94bW"
-}
