--- src/appmixer/pinterest/core/GetaBoard/component.json
@@ -1,66 +1,0 @@
-{
-    "name": "appmixer.pinterest.core.GetaBoard",
-    "author": "Appmixer <info@appmixer.com>",
-    "description": "Retrieves information about a specific board.",
-    "version": "1.0.0",
-    "private": false,
-    "auth": {
-        "service": "appmixer:pinterest"
-    },
-    "quota": {
-        "manager": "appmixer:pinterest",
-        "resources": "requests"
-    },
-    "inPorts": [
-        {
-            "name": "in",
-            "schema": {
-                "type": "object",
-                "properties": {
-                    "board_id": {
-                        "type": "string"
-                    }
-                }
-            },
-            "inspector": {
-                "inputs": {
-                    "board_id": {
-                        "type": "text",
-                        "tooltip": "The ID of the board to retrieve.",
-                        "label": "Board Id",
-                        "index": 0
-                    }
-                }
-            }
-        }
-    ],
-    "outPorts": [
-        {
-            "name": "out",
-            "options": [
-                {
-                    "label": "Id",
-                    "value": "id",
-                    "schema": {
-                        "type": "string"
-                    }
-                },
-                {
-                    "label": "Name",
-                    "value": "name",
-                    "schema": {
-                        "type": "string"
-                    }
-                },
-                {
-                    "label": "Description",
-                    "value": "description",
-                    "schema": {
-                        "type": "string"
-                    }
-                }
-            ]
-        }
-    ],
-    "icon": "data:image/svg+xml;base64,PD94bW"
-}
