--- src/appmixer/pinterest/core/UpdateaBoard/component.json
@@ -1,84 +1,0 @@
-{
-    "name": "appmixer.pinterest.core.UpdateaBoard",
-    "author": "Appmixer <info@appmixer.com>",
-    "description": "Updates an existing board with new information.",
-    "version": "1.0.0",
-    "private": false,
-    "auth": {
-        "service": "appmixer:pinterest"
-    },
-    "quota": {
-        "manager": "appmixer:pinterest",
-        "resources": "requests"
-    },
-    "inPorts": [
-        {
-            "name": "in",
-            "schema": {
-                "type": "object",
-                "properties": {
-                    "board_id": {
-                        "type": "string"
-                    },
-                    "name": {
-                        "type": "string"
-                    },
-                    "description": {
-                        "type": "string"
-                    }
-                }
-            },
-            "inspector": {
-                "inputs": {
-                    "board_id": {
-                        "type": "text",
-                        "tooltip": "The ID of the board to update.",
-                        "label": "Board Id",
-                        "index": 0
-                    },
-                    "name": {
-                        "type": "text",
-                        "tooltip": "The new name for the board.",
-                        "label": "Name",
-                        "index": 1
-                    },
-                    "description": {
-                        "type": "text",
-                        "tooltip": "The new description for the board.",
-                        "label": "Description",
-                        "index": 2
-                    }
-                }
-            }
-        }
-    ],
-    "outPorts": [
-        {
-            "name": "out",
-            "options": [
-                {
-                    "label": "Id",
-                    "value": "id",
-                    "schema": {
-                        "type": "string"
-                    }
-                },
-                {
-                    "label": "Name",
-                    "value": "name",
-                    "schema": {
-                        "type": "string"
-                    }
-                },
-                {
-                    "label": "Description",
-                    "value": "description",
-                    "schema": {
-                        "type": "string"
-                    }
-                }
-            ]
-        }
-    ],
-    "icon": "data:image/svg+xml;base64,PD94bW"
-}
